Bug 2048352 - Split the push workflow into VCS agnostic base classes
Extract the version control agnostic parts of the mercurial push workflow
into BaseRepository and BaseWorker (new vcs module): building and applying a
stack of Phabricator patches, committing try_task_config.json, the skippable
files detection, the push retry loop and the worker output handling.

The Mercurial specific parts stay in MercurialRepository (renamed from
Repository to avoid confusion with the upcoming Git implementation) and
MercurialWorker: robustcheckout and batch clones, revision lookups with the
Lando git2hg fallback, patch application through hg import, ssh push and the
treestatus wait.

No behavior change: this prepares the introduction of a Git implementation
of the same workflow as subclasses of the new bases.
