+++ bot/code_review_bot/mercurial.py
@@ -5,7 +5,6 @@
 import atexit
 import fcntl
 import io
-import json
 import os
 import tempfile
 import time
@@ -13,33 +12,19 @@ from datetime import datetime
 
 import hglib
 import requests
-import rs_parsepatch
 import structlog
 from libmozdata.lando import LandoCommitMapAPI, LandoMissingCommit
 from libmozdata.phabricator import PhabricatorPatch
 
-from code_review_bot.sources.phabricator import PhabricatorBuild
+from code_review_bot.vcs import BaseRepository, BaseWorker
 
 logger = structlog.get_logger(__name__)
 
-TREEHERDER_URL = "https://treeherder.mozilla.org/#/jobs?repo={}&revision={}"
 DEFAULT_AUTHOR = "code review bot <release-mgmt-analysis@mozilla.com>"
 # On build failure, check try status until available every 5 minutes and up to 24h
 TRY_STATUS_URL = "https://treestatus.prod.lando.prod.cloudops.mozgcp.net/trees/try"
 TRY_STATUS_DELAY = 5 * 60
 TRY_STATUS_MAX_WAIT = 24 * 60 * 60
-# Number of allowed retries on an unexpected push fail
-MAX_PUSH_RETRIES = 4
-# Wait successive exponential delays: 6sec, 36sec, 3.6min, 21.6min
-PUSH_RETRY_EXPONENTIAL_DELAY = 6
-
-logger = structlog.get_logger(__name__)
-
-
-class RetryNeeded(Exception):
-    """
-    Raised when retrying a mercurial build is needed
-    """
 
 
 def hg_run(cmd):
@@ -146,31 +131,22 @@ def robust_checkout(
     hg_run(cmd)
 
 
-class Repository:
+class MercurialRepository(BaseRepository):
     """
     A Mercurial repository with its try server credentials
     """
 
+    DEFAULT_REVISION = "default"
+
     def __init__(self, config, cache_root):
-        assert isinstance(config, dict)
-        self.name = config["name"]
-        self.url = config["url"]
-        self.dir = os.path.join(cache_root, config["name"])
+        super().__init__(config, cache_root)
         self.share_base_dir = os.path.join(cache_root, f"{config['name']}-shared")
         self.checkout_mode = config.get("checkout", "batch")
         self.batch_size = config.get("batch_size", 10000)
-        self.try_url = config["try_url"]
-        self.try_name = config.get("try_name", "try")
-        self.default_revision = config.get("default_revision", "default")
-
-        # Apply patches to the latest revision when `True`.
-        self.use_latest_revision = config.get("use_latest_revision", False)
 
         # Crash when configuration requests try syntax
         if config.get("try_mode") == "syntax":
             raise Exception("Try syntax mode is deprecated")
-
-        self._repo = None
 
         # Write ssh key from secret
         _, self.ssh_key_path = tempfile.mkstemp(suffix=".key")
@@ -189,9 +165,6 @@ class Repository:
 
         # Remove key when finished
         atexit.register(self.end_of_life)
-
-    def __str__(self):
-        return self.name
 
     def end_of_life(self):
         os.unlink(self.ssh_key_path)
@@ -276,57 +249,11 @@ class Repository:
             # Base revision may reference a Git hash on new repositories
             return self.get_mercurial_base_hash(base_rev_hash)
 
-    def apply_build(self, build):
-        """
-        Apply a stack of patches to mercurial repo
-        and commit them one by one
-        """
-        assert isinstance(build, PhabricatorBuild)
-        assert len(build.stack) > 0, "No patches to apply"
-        assert all(map(lambda p: isinstance(p, PhabricatorPatch), build.stack))
-
-        # Find the first unknown base revision
-        needed_stack = []
-        for patch in reversed(build.stack):
-            # Skip already merged patches
-            if patch.merged:
-                logger.info(
-                    f"Skip applying patch {patch.id} as it's already been merged upstream"
-                )
-                continue
-
-            # Add the patch into the stack only if not already merged !
-            needed_stack.insert(0, patch)
-
-            # Stop as soon as a base revision is available
-            if self.has_revision(patch.base_revision):
-                logger.info(f"Stopping at revision {patch.base_revision}")
-                break
-
-        if not needed_stack:
-            logger.info("All the patches are already applied")
-            return
-
-        hg_base = self.get_base_identifier(needed_stack)
-
-        # When base revision is missing, update to default revision
-        build.base_revision = hg_base
-        build.missing_base_revision = not self.has_revision(hg_base)
-        if build.missing_base_revision:
-            logger.warning(
-                "Missing base revision from Phabricator",
-                revision=hg_base,
-                fallback=self.default_revision,
-            )
-            hg_base = self.default_revision
-
-        # Store the actual base revision we used
-        build.actual_base_revision = hg_base
-
-        # Update the repo to base revision
+    def checkout_base(self, base):
+        """Update the local checkout to the base revision"""
         try:
-            logger.info(f"Updating repo to revision {hg_base}")
-            self.repo.update(rev=hg_base, clean=True)
+            logger.info(f"Updating repo to revision {base}")
+            self.repo.update(rev=base, clean=True)
 
             # See if the repo is clean
             repo_status = self.repo.status(
@@ -335,111 +262,79 @@ class Repository:
             if len(repo_status) != 0:
                 logger.warn(
                     "Repo is dirty!",
-                    revision=hg_base,
+                    revision=base,
                     repo=self.name,
                     repo_status=repo_status,
                 )
 
         except hglib.error.CommandError:
-            raise Exception(f"Failed to update to revision {hg_base}")
+            raise Exception(f"Failed to update to revision {base}")
 
-        # In this case revision is `hg_base`
-        logger.info("Updated repo", revision=hg_base, repo=self.name)
+        # In this case revision is `base`
+        logger.info("Updated repo", revision=base, repo=self.name)
 
-        def get_author(commit):
-            """Helper to build a mercurial author from Phabricator data"""
-            author = commit.get("author")
-            if author is None:
-                return DEFAULT_AUTHOR
-            if author["name"] and author["email"]:
-                # Build clean version without quotes
-                return f"{author['name']} <{author['email']}>"
-            return author["raw"]
+    @staticmethod
+    def get_author(commit):
+        """Helper to build a mercurial author from Phabricator data"""
+        author = commit.get("author")
+        if author is None:
+            return DEFAULT_AUTHOR
+        if author["name"] and author["email"]:
+            # Build clean version without quotes
+            return f"{author['name']} <{author['email']}>"
+        return author["raw"]
 
-        for patch in needed_stack:
-            if patch.commits:
-                # Use the first commit only
-                commit = patch.commits[0]
-                message = "{}\n".format(commit["message"])
-                user = get_author(commit)
-            else:
-                # We should always have some commits here
-                logger.warning("Missing commit on patch", id=patch.id)
-                message = ""
-                user = DEFAULT_AUTHOR
-            message += f"Differential Diff: {patch.phid}"
+    def apply_patch(self, patch, message, commit):
+        """Apply a single patch on the local checkout and commit it"""
+        user = self.get_author(commit) if commit else DEFAULT_AUTHOR
 
-            logger.info("Applying patch", phid=patch.phid, message=message)
-            patches = io.BytesIO(patch.patch.encode("utf-8"))
-            try:
-                self.repo.import_(
-                    patches=patches,
+        patches = io.BytesIO(patch.patch.encode("utf-8"))
+        try:
+            self.repo.import_(
+                patches=patches,
+                message=message.encode("utf-8"),
+                user=user.encode("utf-8"),
+                similarity=95,
+            )
+        except hglib.error.CommandError as e:
+            logger.warning(
+                (
+                    f"Mercurial command from hglib failed: {e}. "
+                    "Retrying with --config ui.patch=patch."
+                ),
+                phid=patch.phid,
+                exc_info=True,
+            )
+            patches.seek(0)
+            # Same method as repo.import_() but with the extra argument "--config ui.patch=patch".
+            # https://repo.mercurial-scm.org/python-hglib/file/484b56ac4aec/hglib/client.py#l959
+            self.repo.rawcommand(
+                hglib.util.cmdbuilder(
+                    b"import",
                     message=message.encode("utf-8"),
                     user=user.encode("utf-8"),
                     similarity=95,
+                    config="ui.patch=patch",
+                    *patches,
                 )
-            except hglib.error.CommandError as e:
-                logger.warning(
-                    (
-                        f"Mercurial command from hglib failed: {e}. "
-                        "Retrying with --config ui.patch=patch."
-                    ),
-                    phid=patch.phid,
-                    exc_info=True,
+            )
+            # When using an external patch util mercurial won't automatically handle add/remove/renames
+            self.repo.rawcommand(
+                hglib.util.cmdbuilder(
+                    b"addremove",
+                    similarity=95,
                 )
-                patches.seek(0)
-                # Same method as repo.import_() but with the extra argument "--config ui.patch=patch".
-                # https://repo.mercurial-scm.org/python-hglib/file/484b56ac4aec/hglib/client.py#l959
-                self.repo.rawcommand(
-                    hglib.util.cmdbuilder(
-                        b"import",
-                        message=message.encode("utf-8"),
-                        user=user.encode("utf-8"),
-                        similarity=95,
-                        config="ui.patch=patch",
-                        *patches,
-                    )
-                )
-                # When using an external patch util mercurial won't automatically handle add/remove/renames
-                self.repo.rawcommand(
-                    hglib.util.cmdbuilder(
-                        b"addremove",
-                        similarity=95,
-                    )
-                )
-            except Exception as e:
-                logger.info(
-                    f"Failed to apply patch: {e}",
-                    phid=patch.phid,
-                    exc_info=True,
-                )
-                raise
+            )
+        except Exception as e:
+            logger.info(
+                f"Failed to apply patch: {e}",
+                phid=patch.phid,
+                exc_info=True,
+            )
+            raise
 
-    def add_try_commit(self, build):
-        """
-        Build and commit the file configuring try
-        with try_task_config.json and the code-review workflow parameters in JSON
-        """
-        path = os.path.join(self.dir, "try_task_config.json")
-        config = {
-            "version": 2,
-            "parameters": {
-                "target_tasks_method": "codereview",
-                "optimize_target_tasks": True,
-                "phabricator_diff": build.target_phid,
-            },
-        }
-        diff_phid = build.stack[-1].phid
-
-        if build.revision_url:
-            message = f"try_task_config for {build.revision_url}"
-        else:
-            message = "try_task_config for code-review"
-        message += f"\nDifferential Diff: {diff_phid}"
-
-        # Write content as json and commit it
-        with open(path, "w") as f:
-            json.dump(config, f, sort_keys=True, indent=4)
+    def commit_try_task_config(self, path, message):
+        """Commit the try_task_config.json file as the bot"""
         self.repo.add(path.encode("utf-8"))
         self.repo.commit(message=message, user=DEFAULT_AUTHOR)
 
@@ -456,6 +351,10 @@ class Repository:
             force=True,
         )
         return tip
+
+    def revision_id(self, tip):
+        """Extract the revision identifier from a push_to_try result"""
+        return tip.node.decode("utf-8")
 
     def clean(self):
         """
@@ -478,10 +377,15 @@ class Repository:
         self.repo.pull()
 
 
-class MercurialWorker:
+class MercurialWorker(BaseWorker):
     """
     Mercurial worker maintaining several local clones
     """
+
+    VCS_ERROR = hglib.error.CommandError
+    VCS_NAME = "Mercurial"
+    FAILURE_MODE = "fail:mercurial"
+    REPOSITORY_CLASS = MercurialRepository
 
     ELIGIBLE_RETRY_ERRORS = [
         error.lower()
@@ -491,67 +395,6 @@ class MercurialWorker:
             "error: EOF occurred in violation of protocol",
         ]
     ]
-
-    def __init__(
-        self,
-        skippable_files=[],
-    ):
-        self.skippable_files = skippable_files
-
-    def run(self, repository, build):
-        """
-        Apply the stack of patches from the build, handling retries
-        in case of try server errors
-        """
-        while build.retries <= MAX_PUSH_RETRIES:
-            start = time.time()
-
-            if build.retries:
-                logger.warning(
-                    "Trying to apply build's diff after a remote push error "
-                    f"[{build.retries}/{MAX_PUSH_RETRIES}]"
-                )
-
-            try:
-                return self.handle_build(repository, build)
-            except RetryNeeded:
-                build.retries += 1
-
-                if build.retries > MAX_PUSH_RETRIES:
-                    error_log = "Max number of retries has been reached pushing the build to try repository"
-                    logger.warn("Mercurial error on diff", error=error_log, build=build)
-                    return (
-                        "fail:mercurial",
-                        build,
-                        {"message": error_log, "duration": time.time() - start},
-                    )
-
-                # Ensure try is opened
-                self.wait_try_available()
-
-                # Wait an exponential time before retrying the build
-                delay = PUSH_RETRY_EXPONENTIAL_DELAY**build.retries
-                logger.info(
-                    f"An error occurred pushing the build to try, retrying after {delay}s"
-                )
-                time.sleep(delay)
-
-    def is_commit_skippable(self, build):
-        def get_files_touched_in_diff(rawdiff):
-            patched = []
-            for parsed_diff in rs_parsepatch.get_diffs(rawdiff):
-                # filename is sometimes of format 'test.txt  Tue Feb 05 17:23:40 2019 +0100'
-                # fix after https://github.com/mozilla/rust-parsepatch/issues/61
-                if "filename" in parsed_diff:
-                    filename = parsed_diff["filename"].split(" ")[0]
-                    patched.append(filename)
-            return patched
-
-        return any(
-            patched_file in self.skippable_files
-            for rev in build.stack
-            for patched_file in get_files_touched_in_diff(rev.patch)
-        )
 
     def wait_try_available(self):
         """
@@ -582,84 +425,9 @@ class MercurialWorker:
             )
             time.sleep(TRY_STATUS_DELAY)
 
-    def is_eligible_for_retry(self, error):
-        """
-        Given a Mercurial error message, if it's an error likely due to a
-        temporary connection problem, consider it as eligible for retry.
-        """
-        error = error.lower()
-        return any(
-            eligible_message in error for eligible_message in self.ELIGIBLE_RETRY_ERRORS
-        )
-
-    def handle_build(self, repository, build):
-        """
-        Try to load and apply a diff on local clone
-        If successful, push to try and send a treeherder link
-        In case of an unexpected push failure, retry up to MAX_PUSH_RETRIES
-        times by putting the build task back in the queue
-
-        If the build fail, send a unit result with a warning message
-        """
-        assert isinstance(repository, Repository)
-        start = time.time()
-
-        try:
-            # Start by cleaning the repo
-            repository.clean()
-
-            # First apply patches on local repo
-            repository.apply_build(build)
-
-            # Check Eligibility: some commits don't need to be pushed to try.
-            if self.is_commit_skippable(build):
-                logger.info("This patch series is ineligible for automated try push")
-                return (
-                    "fail:ineligible",
-                    build,
-                    {
-                        "message": "Modified files match skippable internal configuration files",
-                        "duration": time.time() - start,
-                    },
-                )
-
-            # Configure the try task
-            repository.add_try_commit(build)
-
-            # Then push that stack on try
-            tip = repository.push_to_try()
-            logger.info("Diff has been pushed !")
-
-            # Publish Treeherder link
-            uri = TREEHERDER_URL.format(repository.try_name, tip.node.decode("utf-8"))
-        except hglib.error.CommandError as e:
-            # Format nicely the error log
-            error_log = e.err
-            if isinstance(error_log, bytes):
-                error_log = error_log.decode("utf-8")
-
-            if self.is_eligible_for_retry(error_log):
-                raise RetryNeeded
-
-            logger.warn(
-                "Mercurial error on diff", error=error_log, args=e.args, build=build
-            )
-            return (
-                "fail:mercurial",
-                build,
-                {"message": error_log, "duration": time.time() - start},
-            )
-
-        except Exception as e:
-            logger.warn("Failed to process diff", error=e, build=build)
-            return (
-                "fail:general",
-                build,
-                {"message": str(e), "duration": time.time() - start},
-            )
-
-        return (
-            "success",
-            build,
-            {"treeherder_url": uri, "revision": tip.node.decode("utf-8")},
-        )
+    def format_error(self, error):
+        """Extract a readable error log from a Mercurial exception"""
+        error_log = error.err
+        if isinstance(error_log, bytes):
+            error_log = error_log.decode("utf-8")
+        return error_log
